# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inventory API",
    "version": "1.0.0",
    "description": "API сервісу інвентаризації"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Реєстрація нового пристрою",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "inventory_name": { "type": "string" },
                  "description": { "type": "string" },
                  "photo": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/inventory": {
      "get": {
        "summary": "Отримати список всіх речей",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/inventory/{id}": {
      "get": {
        "summary": "Отримати річ за ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "summary": "Оновити дані речі",
        "parameters": [
           { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inventory_name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
            "200": { "description": "Updated" },
            "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "summary": "Видалити річ",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not Found" }
        }
      }
    }
  }
}